Fix testing, fix graphite plugin recipe bugs, cleanup (+12 squashed commits) Squashed commits: [10a37f9] Cleaner attribute listing [269f9e7] Moving the plugin packages out of attributes and into recipe
Easier to test this part of the recipe if implemented this way. Chefspec/rspec was having a hard time rendering what packages were needed for the install, this took out the abstraction and allowed rspec to just work..
[992224f] Lacking consistency
[ac9ce9d] Adding explicit action for rspec testing
[ddb614e] Adding rspec documentation and color output

Make it much easier to parse the test output to see what is going wrong and where.
[442968c] Adding supported platforms definition to call in tests

Saw this implementation method in the apache2 cookbook, very useful to quickly add testing for all your supported platforms. See default_spec test for a test implementation example
[28ae93e] Lacking consistency
[19dcff6] Completely reworked, added more tests as well
[ec58006] Adding new spec test for the graphite plugin recipe
[8986ec7] This test never really worked right, removed
[9245fa9] Condense tests, adding testing for the graphite plugin

Includes test for the files being pulled down for the plugin, content of files and the cronjob being added
[dbee4c0] CHANGELOG Update/Cleanup

 * Standardize markdown formatting current and past entries
 * Add new content related to release 3.0.3