Updating the graphite plugin cronjob
After some thought I have decided it is better to leave as many defaults so it will run as root crontab and run minutely still.. It is extra data but if you have running jobs that you want to track stats for, this will add more data to analyze. I am also adding a new attribute that will have to be defined in order for the cronjob to be turned on. This allows some additional flexibility in case a user doesn't like my cron settings. Also avoids temptation to do a rewind someplace..